+++ test/unit/fuzz-tools.test.js
@@ -314,6 +314,7 @@ import * as kudoV334 from '../../lib/kudo-v334.js';
 import * as niceV335 from '../../lib/nice-v335.js';
 import * as jnetV336 from '../../lib/jnet-v336.js';
 import * as outerbridgeV337 from '../../lib/outerbridge-v337.js';
+import * as icrsV338 from '../../lib/icrs-v338.js';
 import * as tbTesting from '../../lib/tb-testing.js';
 
 const MODULES = {
@@ -573,6 +574,7 @@ const MODULES = {
   'nice-v335.js': niceV335,
   'jnet-v336.js': jnetV336,
   'outerbridge-v337.js': outerbridgeV337,
+  'icrs-v338.js': icrsV338,
   'tb-testing.js': tbTesting,
 };